Change the name to plot_transports_prix : it gives 1) the evolution of the price of fuel 2) the evolution of the implied tax rate.
